separate multi-cluster checks and run after add-ons
