+++ pkg/healthcheck/healthcheck.go
@@ -25,7 +25,6 @@ import (
 	log "github.com/sirupsen/logrus"
 	appsv1 "k8s.io/api/apps/v1"
 	corev1 "k8s.io/api/core/v1"
-	v1 "k8s.io/api/rbac/v1"
 	kerrors "k8s.io/apimachinery/pkg/api/errors"
 	"k8s.io/apimachinery/pkg/api/meta"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
@@ -35,7 +34,6 @@ import (
 	yamlDecoder "k8s.io/apimachinery/pkg/util/yaml"
 	k8sVersion "k8s.io/apimachinery/pkg/version"
 	"k8s.io/client-go/kubernetes"
-	"k8s.io/client-go/tools/clientcmd"
 	apiregistrationv1client "k8s.io/kube-aggregator/pkg/client/clientset_generated/clientset/typed/apiregistration/v1"
 	"sigs.k8s.io/yaml"
 )
@@ -139,22 +137,15 @@ const (
 	/// plugin is installed and ready
 	LinkerdCNIPluginChecks CategoryID = "linkerd-cni-plugin"
 
-	// LinkerdMulticlusterChecks adds a series of checks to validate
-	// that the multicluster setup is working as expected
-	LinkerdMulticlusterChecks CategoryID = "linkerd-multicluster"
-
 	// LinkerdCNIResourceLabel is the label key that is used to identify
 	// whether a Kubernetes resource is related to the install-cni command
 	// The value is expected to be "true", "false" or "", where "false" and
 	// "" are equal, making "false" the default
 	LinkerdCNIResourceLabel = "linkerd.io/cni-resource"
 
-	linkerdCNIDisabledSkipReason        = "skipping check because CNI is not enabled"
-	linkerdCNIResourceName              = "linkerd-cni"
-	linkerdCNIConfigMapName             = "linkerd-cni-config"
-	linkerdServiceMirrorComponentName   = "linkerd-service-mirror"
-	linkerdServiceMirrorClusterRoleName = "linkerd-service-mirror-access-local-resources"
-	linkerdServiceMirrorRoleName        = "linkerd-service-mirror-read-remote-creds"
+	linkerdCNIDisabledSkipReason = "skipping check because CNI is not enabled"
+	linkerdCNIResourceName       = "linkerd-cni"
+	linkerdCNIConfigMapName      = "linkerd-cni-config"
 
 	// linkerdTapAPIServiceName is the name of the tap api service
 	// This key is passed to checkApiSercice method to check whether
@@ -201,30 +192,6 @@ var ExpectedServiceAccountNames = []string{
 type expectedPolicy struct {
 	resources []string
 	verbs     []string
-}
-
-var expectedServiceMirrorClusterRolePolicies = []expectedPolicy{
-	{
-		resources: []string{"endpoints", "services"},
-		verbs:     []string{"list", "get", "watch", "create", "delete", "update"},
-	},
-	{
-		resources: []string{"namespaces"},
-		verbs:     []string{"create", "list", "get", "watch"},
-	},
-}
-
-var expectedServiceMirrorRolePolicies = []expectedPolicy{
-	{
-		resources: []string{"secrets"},
-		verbs:     []string{"list", "get", "watch"},
-	},
-}
-
-var expectedServiceMirrorRemoteClusterPolicyVerbs = []string{
-	"get",
-	"list",
-	"watch",
 }
 
 var (
@@ -399,6 +366,7 @@ func NewHealthChecker(categoryIDs []CategoryID, options *Options) *HealthChecker
 	}
 
 	hc.categories = append(hc.allCategories(), hc.addOnCategories()...)
+	hc.categories = append(hc.categories, hc.multiClusterCategory())
 
 	checkMap := map[CategoryID]struct{}{}
 	for _, category := range categoryIDs {
@@ -1250,97 +1218,6 @@ func (hc *HealthChecker) allCategories() []category {
 				},
 			},
 		},
-		{
-			id: LinkerdMulticlusterChecks,
-			checkers: []checker{
-				{
-					description: "service mirror controller is running",
-					hintAnchor:  "l5d-multicluster-service-mirror-running",
-					fatal:       true,
-					check: func(context.Context) error {
-						return hc.checkServiceMirrorController()
-					},
-				},
-				{
-					description: "service mirror controller ClusterRoles exist",
-					hintAnchor:  "l5d-multicluster-cluster-role-exist",
-					check: func(context.Context) error {
-						if hc.Options.ShouldCheckMulticluster {
-							return hc.checkClusterRoles(true, []string{linkerdServiceMirrorClusterRoleName}, hc.serviceMirrorComponentsSelector())
-						}
-						return &SkipError{Reason: "not checking muticluster"}
-					},
-				},
-				{
-					description: "service mirror controller ClusterRoleBindings exist",
-					hintAnchor:  "l5d-multicluster-cluster-role-binding-exist",
-					check: func(context.Context) error {
-						if hc.Options.ShouldCheckMulticluster {
-							return hc.checkClusterRoleBindings(true, []string{linkerdServiceMirrorClusterRoleName}, hc.serviceMirrorComponentsSelector())
-						}
-						return &SkipError{Reason: "not checking muticluster"}
-					},
-				},
-				{
-					description: "service mirror controller Roles exist",
-					hintAnchor:  "l5d-multicluster-role-exist",
-					check: func(context.Context) error {
-						if hc.Options.ShouldCheckMulticluster {
-							return hc.checkRoles(true, hc.serviceMirrorNs, []string{linkerdServiceMirrorRoleName}, hc.serviceMirrorComponentsSelector())
-						}
-						return &SkipError{Reason: "not checking muticluster"}
-					},
-				},
-				{
-					description: "service mirror controller RoleBindings exist",
-					hintAnchor:  "l5d-multicluster-role-binding-exist",
-					check: func(context.Context) error {
-						if hc.Options.ShouldCheckMulticluster {
-							return hc.checkRoleBindings(true, hc.serviceMirrorNs, []string{linkerdServiceMirrorRoleName}, hc.serviceMirrorComponentsSelector())
-						}
-						return &SkipError{Reason: "not checking muticluster"}
-					},
-				},
-				{
-					description: "service mirror controller ServiceAccounts exist",
-					hintAnchor:  "l5d-multicluster-service-account-exist",
-					check: func(context.Context) error {
-						if hc.Options.ShouldCheckMulticluster {
-							return hc.checkServiceAccounts([]string{linkerdServiceMirrorComponentName}, hc.serviceMirrorNs, hc.serviceMirrorComponentsSelector())
-						}
-						return &SkipError{Reason: "not checking muticluster"}
-					},
-				},
-				{
-					description: "service mirror controller has required permissions",
-					hintAnchor:  "l5d-multicluster-local-rbac-correct",
-					check: func(context.Context) error {
-						return hc.checkServiceMirrorLocalRBAC()
-					},
-				},
-				{
-					description: "service mirror controller can access remote clusters",
-					hintAnchor:  "l5d-smc-remote-remote-clusters-access",
-					check: func(context.Context) error {
-						return hc.checkRemoteClusterConnectivity()
-					},
-				},
-				{
-					description: "all remote cluster gateways are alive",
-					hintAnchor:  "l5d-multicluster-remote-gateways-alive",
-					check: func(ctx context.Context) error {
-						return hc.checkRemoteClusterGatewaysHealth(ctx)
-					},
-				},
-				{
-					description: "clusters share trust anchors",
-					hintAnchor:  "l5d-multicluster-clusters-share-anchors",
-					check: func(ctx context.Context) error {
-						return hc.checkRemoteClusterAnchors()
-					},
-				},
-			},
-		},
 	}
 }
 
@@ -1517,10 +1394,6 @@ func (hc *HealthChecker) controlPlaneComponentsSelector() string {
 	return fmt.Sprintf("%s,!%s", k8s.ControllerNSLabel, LinkerdCNIResourceLabel)
 }
 
-func (hc *HealthChecker) serviceMirrorComponentsSelector() string {
-	return fmt.Sprintf("%s=%s", k8s.ControllerComponentLabel, linkerdServiceMirrorComponentName)
-}
-
 // PublicAPIClient returns a fully configured public API client. This client is
 // only configured if the KubernetesAPIChecks and LinkerdAPIChecks are
 // configured and run first.
@@ -1597,271 +1470,6 @@ func FetchLinkerdConfigMap(k kubernetes.Interface, controlPlaneNamespace string)
 	}
 
 	return cm, configPB, nil
-}
-
-func (hc *HealthChecker) checkServiceMirrorController() error {
-	options := metav1.ListOptions{
-		LabelSelector: hc.serviceMirrorComponentsSelector(),
-	}
-	result, err := hc.kubeAPI.AppsV1().Deployments(corev1.NamespaceAll).List(options)
-	if err != nil {
-		return err
-	}
-
-	// if we have explicitly requested for multicluster to be checked, error out
-	if len(result.Items) == 0 && hc.Options.ShouldCheckMulticluster {
-		return errors.New("Service mirror controller is not present")
-	}
-
-	if len(result.Items) > 0 {
-		hc.Options.ShouldCheckMulticluster = true
-
-		if len(result.Items) > 1 {
-			var errors []string
-			for _, smc := range result.Items {
-				errors = append(errors, fmt.Sprintf("%s/%s", smc.Namespace, smc.Name))
-			}
-			return fmt.Errorf("There are more than one service mirror controllers:\n\t%s", strings.Join(errors, "\n\t"))
-		}
-
-		controller := result.Items[0]
-		if controller.Status.AvailableReplicas < 1 {
-			return fmt.Errorf("Service mirror controller is not available: %s/%s", controller.Namespace, controller.Name)
-		}
-		hc.serviceMirrorNs = controller.Namespace
-		return nil
-	}
-
-	return &SkipError{Reason: "not checking muticluster"}
-}
-
-func comparePermissions(expected, actual []string) error {
-	sort.Strings(expected)
-	sort.Strings(actual)
-
-	expectedStr := strings.Join(expected, ",")
-	actualStr := strings.Join(actual, ",")
-
-	if expectedStr != actualStr {
-		return fmt.Errorf("expected %s, got %s", expectedStr, actualStr)
-	}
-
-	return nil
-}
-
-func verifyRule(expected expectedPolicy, actual []v1.PolicyRule) error {
-	for _, rule := range actual {
-		if err := comparePermissions(expected.resources, rule.Resources); err == nil {
-			if err := comparePermissions(expected.verbs, rule.Verbs); err != nil {
-				return fmt.Errorf("unexpected verbs %s", err)
-			}
-			return nil
-		}
-	}
-	return fmt.Errorf("could not fine rule for %s", strings.Join(expected.resources, ","))
-}
-
-func (hc *HealthChecker) checkServiceMirrorLocalRBAC() error {
-	if hc.Options.ShouldCheckMulticluster {
-		var errors []string
-
-		clusterRole, err := hc.kubeAPI.RbacV1().ClusterRoles().Get(linkerdServiceMirrorClusterRoleName, metav1.GetOptions{})
-		if err != nil {
-			return fmt.Errorf("Could not obtain service mirror ClusterRole %s: %s", linkerdServiceMirrorClusterRoleName, err)
-		}
-
-		role, err := hc.kubeAPI.RbacV1().Roles(hc.serviceMirrorNs).Get(linkerdServiceMirrorRoleName, metav1.GetOptions{})
-		if err != nil {
-			return fmt.Errorf("Could not obtain service mirror Role %s : %s", linkerdServiceMirrorRoleName, err)
-		}
-
-		if len(clusterRole.Rules) != len(expectedServiceMirrorClusterRolePolicies) {
-			return fmt.Errorf("Service mirror ClusterRole %s has %d policy rules, expected %d", clusterRole.Name, len(clusterRole.Rules), len(expectedServiceMirrorClusterRolePolicies))
-		}
-
-		for _, rule := range expectedServiceMirrorClusterRolePolicies {
-			if err := verifyRule(rule, clusterRole.Rules); err != nil {
-				errors = append(errors, fmt.Sprintf("Service mirror ClusterRole %s: %s", clusterRole.Name, err))
-			}
-		}
-
-		if len(role.Rules) != len(expectedServiceMirrorRolePolicies) {
-			return fmt.Errorf("Service mirror Role %s has %d policy rules, expected %d", role.Name, len(role.Rules), len(expectedServiceMirrorRolePolicies))
-		}
-
-		for _, rule := range expectedServiceMirrorRolePolicies {
-			if err := verifyRule(rule, role.Rules); err != nil {
-				errors = append(errors, fmt.Sprintf("Service mirror Role %s: %s", role.Name, err))
-			}
-		}
-
-		if len(errors) > 0 {
-			return fmt.Errorf(strings.Join(errors, "\n"))
-		}
-
-		return nil
-	}
-	return &SkipError{Reason: "not checking muticluster"}
-}
-
-func (hc *HealthChecker) checkRemoteClusterAnchors() error {
-	if len(hc.remoteClusterConfigs) == 0 {
-		return &SkipError{Reason: "no remote cluster configs"}
-	}
-
-	localAnchors, err := tls.DecodePEMCertificates(hc.linkerdConfig.Global.IdentityContext.TrustAnchorsPem)
-	if err != nil {
-		return fmt.Errorf("Cannot parse local trust anchors: %s", err)
-	}
-
-	var offendingClusters []string
-	for _, cfg := range hc.remoteClusterConfigs {
-
-		clientConfig, err := clientcmd.RESTConfigFromKubeConfig(cfg.APIConfig)
-		if err != nil {
-			offendingClusters = append(offendingClusters, fmt.Sprintf("* %s: unable to parse api config", cfg.ClusterName))
-			continue
-		}
-
-		remoteAPI, err := k8s.NewAPIForConfig(clientConfig, "", []string{}, requestTimeout)
-		if err != nil {
-			offendingClusters = append(offendingClusters, fmt.Sprintf("* %s: unable to instantiate api", cfg.ClusterName))
-			continue
-		}
-
-		_, cfMap, err := FetchLinkerdConfigMap(remoteAPI, cfg.LinkerdNamespace)
-		if err != nil {
-			offendingClusters = append(offendingClusters, fmt.Sprintf("* %s: unable to fetch anchors: %s", cfg.ClusterName, err))
-			continue
-		}
-		remoteAnchors, err := tls.DecodePEMCertificates(cfMap.Global.IdentityContext.TrustAnchorsPem)
-		if err != nil {
-			offendingClusters = append(offendingClusters, fmt.Sprintf("* %s: cannot parse trust anchors", cfg.ClusterName))
-			continue
-		}
-
-		// we fail early if the lens are not the same. If they are the
-		// same, we can only compare certs one way and be sure we have
-		// identical anchors
-		if len(remoteAnchors) != len(localAnchors) {
-			offendingClusters = append(offendingClusters, fmt.Sprintf("* %s", cfg.ClusterName))
-			continue
-		}
-
-		localAnchorsMap := make(map[string]*x509.Certificate)
-		for _, c := range localAnchors {
-			localAnchorsMap[string(c.Signature)] = c
-		}
-
-		for _, remote := range remoteAnchors {
-			local, ok := localAnchorsMap[string(remote.Signature)]
-			if !ok || !local.Equal(remote) {
-				offendingClusters = append(offendingClusters, fmt.Sprintf("* %s", cfg.ClusterName))
-				break
-			}
-		}
-	}
-
-	if len(offendingClusters) > 0 {
-		return fmt.Errorf("Problematic clusters:\n\t%s", strings.Join(offendingClusters, "\n\t"))
-	}
-
-	return nil
-}
-
-func (hc *HealthChecker) checkRemoteClusterConnectivity() error {
-	if hc.Options.ShouldCheckMulticluster {
-		options := metav1.ListOptions{
-			FieldSelector: fmt.Sprintf("%s=%s", "type", k8s.MirrorSecretType),
-		}
-		secrets, err := hc.kubeAPI.CoreV1().Secrets(corev1.NamespaceAll).List(options)
-		if err != nil {
-			return err
-		}
-
-		if len(secrets.Items) == 0 {
-			return &SkipError{Reason: "no remote cluster configs"}
-		}
-
-		var errors []string
-		for _, s := range secrets.Items {
-			secret := s
-			config, err := sm.ParseRemoteClusterSecret(&secret)
-			if err != nil {
-				errors = append(errors, fmt.Sprintf("*  secret: [%s/%s]: could not parse config secret: %s", secret.Namespace, secret.Name, err))
-				continue
-			}
-
-			clientConfig, err := clientcmd.RESTConfigFromKubeConfig(config.APIConfig)
-			if err != nil {
-				errors = append(errors, fmt.Sprintf("* secret: [%s/%s] cluster: [%s]: unable to parse api config: %s", secret.Namespace, secret.Name, config.ClusterName, err))
-				continue
-			}
-
-			remoteAPI, err := k8s.NewAPIForConfig(clientConfig, "", []string{}, requestTimeout)
-			if err != nil {
-				errors = append(errors, fmt.Sprintf("* secret: [%s/%s] cluster: [%s]: could not instantiate remote api: %s", secret.Namespace, secret.Name, config.ClusterName, err))
-				continue
-			}
-
-			var verbs []string
-			if err := hc.checkCanPerformAction(remoteAPI, "get", corev1.NamespaceAll, "", "v1", "services"); err == nil {
-				verbs = append(verbs, "get")
-			}
-
-			if err := hc.checkCanPerformAction(remoteAPI, "list", corev1.NamespaceAll, "", "v1", "services"); err == nil {
-				verbs = append(verbs, "list")
-			}
-
-			if err := hc.checkCanPerformAction(remoteAPI, "watch", corev1.NamespaceAll, "", "v1", "services"); err == nil {
-				verbs = append(verbs, "watch")
-			}
-
-			if err := comparePermissions(expectedServiceMirrorRemoteClusterPolicyVerbs, verbs); err != nil {
-				errors = append(errors, fmt.Sprintf("* cluster: [%s]: Insufficient Service permissions: %s", config.ClusterName, err))
-			}
-
-			hc.remoteClusterConfigs = append(hc.remoteClusterConfigs, config)
-
-		}
-
-		if len(errors) > 0 {
-			return fmt.Errorf("Problematic clusters:\n\t%s", strings.Join(errors, "\n\t"))
-		}
-		return nil
-	}
-	return &SkipError{Reason: "not checking muticluster"}
-}
-
-func (hc *HealthChecker) checkRemoteClusterGatewaysHealth(ctx context.Context) error {
-	if hc.Options.ShouldCheckMulticluster {
-		if hc.apiClient == nil {
-			return errors.New("public api client uninitialized")
-		}
-		req := &pb.GatewaysRequest{
-			TimeWindow: "1m",
-		}
-		rsp, err := hc.apiClient.Gateways(ctx, req)
-		if err != nil {
-			return err
-		}
-
-		var deadGateways []string
-		if len(rsp.GetOk().GatewaysTable.Rows) == 0 {
-			return &SkipError{Reason: "no remote gateways"}
-		}
-		for _, gtw := range rsp.GetOk().GatewaysTable.Rows {
-			if !gtw.Alive {
-				deadGateways = append(deadGateways, fmt.Sprintf("* cluster: [%s], gateway: [%s/%s]", gtw.ClusterName, gtw.Namespace, gtw.Name))
-			}
-		}
-
-		if len(deadGateways) > 0 {
-			return fmt.Errorf("Some gateways are not alive:\n\t%s", strings.Join(deadGateways, "\n\t"))
-		}
-		return nil
-	}
-	return &SkipError{Reason: "not checking muticluster"}
 }
 
 // checkNamespace checks whether the given namespace exists, and returns an
